requirements.txt
requests>=2.31
beautifulsoup4>=4.12
lxml>=5.0
numpy>=1.24
pandas>=2.0
python-dotenv>=1.2.2